alici_project/requirements.txt
flask
requests
openai
sentence-transformers
python-dotenv
gunicorn
psycopg2-binary